add module file
